% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{db_enable_foreign_key_check}
\alias{db_enable_foreign_key_check}
\alias{db_disable_foreign_key_check}
\title{Enable/disable foreign key constraints in a database}
\usage{
db_enable_foreign_key_check(db)

db_disable_foreign_key_check(db)
}
\arguments{
\item{db}{A DBI connection}
}
\value{
Invisibly returns the response from \link[DBI:dbExecute]{DBI::dbExecute}.
}
\description{
Enable/disable foreign key constraints in a database
}
\section{Functions}{
\itemize{
\item \code{db_disable_foreign_key_check}: Disable foreign key constraints
}}

