% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.R
\name{rel}
\alias{rel}
\alias{new_relational}
\alias{rel_tostring}
\alias{rel_names}
\title{Relational API}
\usage{
new_relational(..., class = NULL)

rel_tostring(rel, ...)

rel_names(rel, ...)
}
\arguments{
\item{...}{Passed on to \code{\link[=structure]{structure()}}}

\item{class}{Classes added in front of the \code{"relational"} base class}
}
\description{
TBD.
}
