% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.R
\name{rel_explain}
\alias{rel_explain}
\title{Print the EXPLAIN output for a DuckDB relation object}
\usage{
rel_explain(rel)
}
\arguments{
\item{rel}{the DuckDB relation object}
}
\description{
Print the EXPLAIN output for a DuckDB relation object
}
\examples{
rel <- rel_from_df(mtcars)
rel_explain(rel)
}
