% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-package.R
\docType{package}
\name{relational-package}
\alias{relational}
\alias{relational-package}
\title{relational: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/defmacro}
  \item \url{https://krlmlr.github.io/defmacro}
  \item Report bugs at \url{https://github.com/krlmlr/defmacro/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

}
\keyword{internal}
