% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{specify_tbl}
\alias{specify_tbl}
\title{Write a table to a database with column and table constraints}
\usage{
specify_tbl(db, df, tbl_name, p_key = NULL, u_keys = NULL,
  nn_keys = NULL, no_null = FALSE, f_keys = NULL)
}
\arguments{
\item{db}{A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object.}

\item{df}{A data frame object.}

\item{tbl_name}{Name of the database.}

\item{p_key}{Column to be constrained as \code{PRIMARY KEY}.}

\item{u_keys}{Character vector with columns to be constrained as \code{UNIQUE}.}

\item{nn_keys}{Character vector with columns to be constrained as \code{NOT NULL}.}

\item{no_null}{Boolean. If \code{TRUE}, constrain all table columsn as \code{NOT NULL}.}

\item{f_keys}{Lists of foreign key declarations.}
}
\description{
Write a table to a database with column and table constraints
}
