% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-package.R
\docType{package}
\name{relational-package}
\alias{relational}
\alias{relational-package}
\title{relational: Relations and Expressions}
\description{
Represent expressions and provide a set of generics for transformations of tables (relations).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/duckdblabs/relational}
  \item \url{https://duckdblabs.github.io/relational}
  \item Report bugs at \url{https://github.com/duckdblabs/relational/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

}
\keyword{internal}
